+++ core/src/test/java/google/registry/persistence/converter/CreateAutoTimestampConverterTest.java
@@ -11,27 +11,27 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import static com.google.common.truth.Truth.assertThat;
 import static google.registry.persistence.transaction.TransactionManagerFactory.jpaTm;
 
+import google.registry.model.CreateAutoTimestamp;
 import google.registry.model.ImmutableObject;
-import google.registry.model.UpdateAutoTimestamp;
-import google.registry.persistence.converter.UpdateAutoTimestampConverter;
 import google.registry.persistence.transaction.JpaTestRules;
 import google.registry.persistence.transaction.JpaTestRules.JpaUnitTestRule;
 import google.registry.testing.FakeClock;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import org.joda.time.DateTime;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link UpdateAutoTimestampConverter}. */
+/** Unit tests for {@link CreateAutoTimestampConverter}. */
 @RunWith(JUnit4.class)
-public class UpdateAutoTimestampConverterTest {
+public class CreateAutoTimestampConverterTest {
 
   private final FakeClock fakeClock = new FakeClock();
 
@@ -44,37 +44,25 @@ public class UpdateAutoTimestampConverterTest {
 
   @Test
   public void testTypeConversion() {
-    TestEntity ent = new TestEntity("myinst", null);
+    CreateAutoTimestamp ts = CreateAutoTimestamp.create(DateTime.parse("2019-09-9T11:39:00Z"));
+    TestEntity ent = new TestEntity("myinst", ts);
+
+    jpaTm().transact(() -> jpaTm().getEntityManager().persist(ent));
+    TestEntity result =
+        jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "myinst"));
+    assertThat(result).isEqualTo(new TestEntity("myinst", ts));
+  }
+
+  @Test
+  public void testAutoInitialization() {
+    CreateAutoTimestamp ts = CreateAutoTimestamp.create(null);
+    TestEntity ent = new TestEntity("autoinit", ts);
 
     jpaTm().transact(() -> jpaTm().getEntityManager().persist(ent));
 
     TestEntity result =
-        jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "myinst"));
-
-    assertThat(result.name).isEqualTo("myinst");
-    assertThat(result.uat.getTimestamp()).isEqualTo(fakeClock.nowUtc());
-  }
-
-  @Test
-  public void testTimeChangesOnSubsequentTransactions() {
-    TestEntity ent1 = new TestEntity("myinst1", null);
-
-    jpaTm().transact(() -> jpaTm().getEntityManager().persist(ent1));
-
-    TestEntity result1 =
-        jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "myinst1"));
-
-    fakeClock.advanceOneMilli();
-
-    TestEntity ent2 = new TestEntity("myinst2", result1.uat);
-
-    jpaTm().transact(() -> jpaTm().getEntityManager().persist(ent2));
-
-    TestEntity result2 =
-        jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "myinst2"));
-
-    assertThat(result1.uat.getTimestamp()).isNotEqualTo(result2.uat.getTimestamp());
-    assertThat(result2.uat.getTimestamp()).isEqualTo(fakeClock.nowUtc());
+        jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "autoinit"));
+    assertThat(result.cat.getTimestamp()).isEqualTo(fakeClock.nowUtc());
   }
 
   @Entity(name = "TestEntity") // Override entity name to avoid the nested class reference.
@@ -82,13 +70,13 @@ public class UpdateAutoTimestampConverterTest {
 
     @Id String name;
 
-    UpdateAutoTimestamp uat;
+    CreateAutoTimestamp cat;
 
     public TestEntity() {}
 
-    TestEntity(String name, UpdateAutoTimestamp uat) {
+    TestEntity(String name, CreateAutoTimestamp cat) {
       this.name = name;
-      this.uat = uat;
+      this.cat = cat;
     }
   }
 }
