+++ core/src/test/java/google/registry/persistence/converter/CurrencyToBillingMapUserTypeTest.java
@@ -12,46 +12,49 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import static com.google.common.truth.Truth.assertThat;
 import static google.registry.persistence.transaction.TransactionManagerFactory.jpaTm;
 
-import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import google.registry.model.ImmutableObject;
-import google.registry.persistence.converter.CidrAddressBlockListConverter;
+import google.registry.model.registrar.Registrar.BillingAccountEntry;
 import google.registry.persistence.transaction.JpaTestRules;
 import google.registry.persistence.transaction.JpaTestRules.JpaUnitTestRule;
-import google.registry.util.CidrAddressBlock;
-import java.util.List;
+import java.util.Map;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import org.hibernate.annotations.Type;
+import org.joda.money.CurrencyUnit;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link CidrAddressBlockListConverter}. */
+/** Unit tests for {@link CurrencyToBillingMapUserType}. */
 @RunWith(JUnit4.class)
-public class CidrAddressBlockListUserTypeTest {
+public class CurrencyToBillingMapUserTypeTest {
   @Rule
   public final JpaUnitTestRule jpaRule =
-      new JpaTestRules.Builder().withEntityClass(TestEntity.class).buildUnitTestRule();
+      new JpaTestRules.Builder()
+          .withInitScript("sql/flyway/V14__load_extension_for_hstore.sql")
+          .withEntityClass(TestEntity.class)
+          .buildUnitTestRule();
 
   @Test
-  public void roundTripConversion_returnsSameCidrAddressBlock() {
-    List<CidrAddressBlock> addresses =
-        ImmutableList.of(
-            CidrAddressBlock.create("0.0.0.0/32"),
-            CidrAddressBlock.create("255.255.255.254/31"),
-            CidrAddressBlock.create("::"),
-            CidrAddressBlock.create("8000::/1"),
-            CidrAddressBlock.create("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/128"));
-    TestEntity testEntity = new TestEntity(addresses);
+  public void roundTripConversion_returnsSameCurrencyToBillingMap() {
+    ImmutableMap<CurrencyUnit, BillingAccountEntry> currencyToBilling =
+        ImmutableMap.of(
+            CurrencyUnit.of("USD"),
+            new BillingAccountEntry(CurrencyUnit.of("USD"), "accountId1"),
+            CurrencyUnit.of("CNY"),
+            new BillingAccountEntry(CurrencyUnit.of("CNY"), "accountId2"));
+    TestEntity testEntity = new TestEntity(currencyToBilling);
     jpaTm().transact(() -> jpaTm().getEntityManager().persist(testEntity));
     TestEntity persisted =
         jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "id"));
-    assertThat(persisted.addresses).isEqualTo(addresses);
+    assertThat(persisted.currencyToBilling).containsExactlyEntriesIn(currencyToBilling);
   }
 
   @Entity(name = "TestEntity") // Override entity name to avoid the nested class reference.
@@ -59,12 +62,13 @@ public class CidrAddressBlockListUserTypeTest {
 
     @Id String name = "id";
 
-    List<CidrAddressBlock> addresses;
+    @Type(type = "google.registry.persistence.converter.CurrencyToBillingMapUserType")
+    Map<CurrencyUnit, BillingAccountEntry> currencyToBilling;
 
     private TestEntity() {}
 
-    private TestEntity(List<CidrAddressBlock> addresses) {
-      this.addresses = addresses;
+    private TestEntity(Map<CurrencyUnit, BillingAccountEntry> currencyToBilling) {
+      this.currencyToBilling = currencyToBilling;
     }
   }
 }
