+++ core/src/test/java/google/registry/persistence/converter/StringListConverterTest.java
@@ -12,62 +12,54 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import static com.google.common.truth.Truth.assertThat;
 import static google.registry.persistence.transaction.TransactionManagerFactory.jpaTm;
 import static org.junit.Assert.assertThrows;
 
-import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableList;
 import google.registry.model.ImmutableObject;
-import google.registry.persistence.converter.MapUserType;
 import google.registry.persistence.transaction.JpaTestRules;
 import google.registry.persistence.transaction.JpaTestRules.JpaUnitTestRule;
-import java.util.Map;
+import java.util.List;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.NoResultException;
-import org.hibernate.annotations.Type;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link MapUserType}. */
+/** Unit tests for {@link StringListConverter}. */
 @RunWith(JUnit4.class)
-public class MapUserTypeTest {
-
-  // Reusing production script sql/flyway/V14__load_extension_for_hstore.sql, which loads the
-  // hstore extension but nothing else.
+public class StringListConverterTest {
   @Rule
   public final JpaUnitTestRule jpaRule =
-      new JpaTestRules.Builder()
-          .withInitScript("sql/flyway/V14__load_extension_for_hstore.sql")
-          .withEntityClass(TestEntity.class)
-          .buildUnitTestRule();
+      new JpaTestRules.Builder().withEntityClass(TestEntity.class).buildUnitTestRule();
 
   @Test
-  public void roundTripConversion_returnsSameMap() {
-    Map<String, String> map = ImmutableMap.of("key1", "value1", "key2", "value2");
-    TestEntity testEntity = new TestEntity(map);
+  public void roundTripConversion_returnsSameStringList() {
+    List<String> tlds = ImmutableList.of("app", "dev", "how");
+    TestEntity testEntity = new TestEntity(tlds);
     jpaTm().transact(() -> jpaTm().getEntityManager().persist(testEntity));
     TestEntity persisted =
         jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "id"));
-    assertThat(persisted.map).containsExactly("key1", "value1", "key2", "value2");
+    assertThat(persisted.tlds).containsExactly("app", "dev", "how");
   }
 
   @Test
   public void testMerge_succeeds() {
-    Map<String, String> map = ImmutableMap.of("key1", "value1", "key2", "value2");
-    TestEntity testEntity = new TestEntity(map);
+    List<String> tlds = ImmutableList.of("app", "dev", "how");
+    TestEntity testEntity = new TestEntity(tlds);
     jpaTm().transact(() -> jpaTm().getEntityManager().persist(testEntity));
     TestEntity persisted =
         jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "id"));
-    persisted.map = ImmutableMap.of("key3", "value3");
+    persisted.tlds = ImmutableList.of("com", "gov");
     jpaTm().transact(() -> jpaTm().getEntityManager().merge(persisted));
     TestEntity updated =
         jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "id"));
-    assertThat(updated.map).containsExactly("key3", "value3");
+    assertThat(updated.tlds).containsExactly("com", "gov");
   }
 
   @Test
@@ -76,45 +68,43 @@ public class MapUserTypeTest {
     jpaTm().transact(() -> jpaTm().getEntityManager().persist(testEntity));
     TestEntity persisted =
         jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "id"));
-    assertThat(persisted.map).isNull();
+    assertThat(persisted.tlds).isNull();
   }
 
   @Test
   public void testEmptyCollection_writesAndReadsEmptyCollectionSuccessfully() {
-    TestEntity testEntity = new TestEntity(ImmutableMap.of());
+    TestEntity testEntity = new TestEntity(ImmutableList.of());
     jpaTm().transact(() -> jpaTm().getEntityManager().persist(testEntity));
     TestEntity persisted =
         jpaTm().transact(() -> jpaTm().getEntityManager().find(TestEntity.class, "id"));
-    assertThat(persisted.map).isEmpty();
+    assertThat(persisted.tlds).isEmpty();
   }
 
   @Test
   public void testNativeQuery_succeeds() throws Exception {
-    executeNativeQuery(
-        "INSERT INTO \"TestEntity\" (name, map) VALUES ('id', 'key1=>value1, key2=>value2')");
+    executeNativeQuery("INSERT INTO \"TestEntity\" (name, tlds) VALUES ('id', '{app, dev}')");
 
     assertThat(
-            getSingleResultFromNativeQuery(
-                "SELECT map -> 'key1' FROM \"TestEntity\" WHERE name = 'id'"))
-        .isEqualTo("value1");
+            getSingleResultFromNativeQuery("SELECT tlds[1] FROM \"TestEntity\" WHERE name = 'id'"))
+        .isEqualTo("app");
     assertThat(
-            getSingleResultFromNativeQuery(
-                "SELECT map -> 'key2' FROM \"TestEntity\" WHERE name = 'id'"))
-        .isEqualTo("value2");
+            getSingleResultFromNativeQuery("SELECT tlds[2] FROM \"TestEntity\" WHERE name = 'id'"))
+        .isEqualTo("dev");
 
-    executeNativeQuery("UPDATE \"TestEntity\" SET map = 'key3=>value3' WHERE name = 'id'");
+    executeNativeQuery("UPDATE \"TestEntity\" SET tlds = '{com, gov}' WHERE name = 'id'");
 
     assertThat(
-            getSingleResultFromNativeQuery(
-                "SELECT map -> 'key3' FROM \"TestEntity\" WHERE name = 'id'"))
-        .isEqualTo("value3");
+            getSingleResultFromNativeQuery("SELECT tlds[1] FROM \"TestEntity\" WHERE name = 'id'"))
+        .isEqualTo("com");
+    assertThat(
+            getSingleResultFromNativeQuery("SELECT tlds[2] FROM \"TestEntity\" WHERE name = 'id'"))
+        .isEqualTo("gov");
 
     executeNativeQuery("DELETE FROM \"TestEntity\" WHERE name = 'id'");
     assertThrows(
         NoResultException.class,
         () ->
-            getSingleResultFromNativeQuery(
-                "SELECT map -> 'key3' FROM \"TestEntity\" WHERE name = 'id'"));
+            getSingleResultFromNativeQuery("SELECT tlds[1] FROM \"TestEntity\" WHERE name = 'id'"));
   }
 
   private static Object getSingleResultFromNativeQuery(String sql) {
@@ -132,13 +122,12 @@ public class MapUserTypeTest {
 
     @Id String name = "id";
 
-    @Type(type = "google.registry.persistence.converter.MapUserType")
-    Map<String, String> map;
+    List<String> tlds;
 
     private TestEntity() {}
 
-    private TestEntity(Map<String, String> map) {
-      this.map = map;
+    private TestEntity(List<String> tlds) {
+      this.tlds = tlds;
     }
   }
 }
